Fix IsValidGridIndex (no walls included) in L and 1x2 rooms

+++ repentogon/LuaInterfaces/Room/LuaRoom.cpp
@@ -189,7 +189,7 @@ LUA_FUNCTION(lua_RoomSetRail) {
 		return luaL_error(L, "Invalid rail type %d\n", rail);
 	}
 
-	room->SetRailType((uint8_t)index, (RailType)rail);
+	room->SetRailType(index, (RailType)rail);
 
 	return 0;
 }
